for norm
