+++ Makefile
@@ -33,6 +33,7 @@ SRCS		+=	$(BUILTIN_DIR)/$(FT_CD_DIR)/canonicalize.c \
 				$(BUILTIN_DIR)/$(FT_CD_DIR)/canonicalize_convert.c \
 				$(BUILTIN_DIR)/$(FT_CD_DIR)/canonicalize_exception.c \
 				$(BUILTIN_DIR)/$(FT_CD_DIR)/canonicalize_sep_join.c \
+				$(BUILTIN_DIR)/$(FT_CD_DIR)/canonicalize_utils.c \
 				$(BUILTIN_DIR)/$(FT_CD_DIR)/chdir.c \
 				$(BUILTIN_DIR)/$(FT_CD_DIR)/check_current_exist.c \
 				$(BUILTIN_DIR)/$(FT_CD_DIR)/check_is_valid_directory.c \
