+++ srcs/builtin/ft_cd/canonicalize_exception.c
@@ -11,10 +11,22 @@ static bool	is_head_path_dot(t_deque *path_elems)
 	return (ft_streq(path_elem, PATH_DOT));
 }
 
+// ./../..  ->  ../..
+static void	erase_head_dot(t_deque *new_path_elems)
+{
+	t_deque_node	*pop_node;
+
+	if (new_path_elems->size >= 2 && is_head_path_dot(new_path_elems))
+	{
+		pop_node = deque_pop_front(new_path_elems);
+		deque_clear_node(&pop_node, del_path_elem);
+	}
+}
+
 // stack
 // ././        -> .
 // .././.././. -> ../..
-// ./../..     -> ../..
+// ./../..     -> ../.. (erase_head_dot)
 void	erase_dot_path_for_relative(t_deque **path_elems)
 {
 	t_deque			*new_path_elems;
@@ -33,11 +45,7 @@ void	erase_dot_path_for_relative(t_deque **path_elems)
 		}
 		deque_add_back(new_path_elems, pop_node);
 	}
-	if (new_path_elems->size >= 2 && is_head_path_dot(new_path_elems))
-	{
-		pop_node = deque_pop_front(new_path_elems);
-		deque_clear_node(&pop_node, del_path_elem);
-	}
+	erase_head_dot(new_path_elems);
 	destroy_path_elems(*path_elems);
 	*path_elems = new_path_elems;
 }
@@ -88,24 +96,4 @@ void	erase_dot_dot_path(t_deque **path_elems)
 	}
 	destroy_path_elems(*path_elems);
 	*path_elems = new_path_elems;
-}
-
-// todo: use ft_strncmp?
-bool	is_head_double_slash(const char *path)
-{
-	return (path[0] == PATH_DELIMITER_CHR \
-			&& path[1] == PATH_DELIMITER_CHR \
-			&& path[2] != PATH_DELIMITER_CHR);
-}
-
-void	handle_double_slash_path(const char *path, char **absolute_path)
-{
-	char	*new_path;
-
-	if (is_head_double_slash(path))
-	{
-		new_path = x_ft_strjoin(PATH_DELIMITER_STR, *absolute_path);
-		ft_free(absolute_path);
-		*absolute_path = new_path;
-	}
 }
